Refatoração de roles e permissions

+++ app/Http/Controllers/ChecklistController.php
@@ -18,6 +18,8 @@ class ChecklistController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Checklist::class);
+
         $message = label_case('Index Checklists ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
         Log::debug($message);
 
@@ -26,29 +28,51 @@ class ChecklistController extends Controller
 
     public function index_data()
     {
+        /*
         if (auth()->user()->isSuperAdmin() || auth()->user()->isAdmin()) {
             $data = Checklist::with('kid')->select('id', 'level', 'situation', 'kid_id', 'created_at');
         } else {
             $data = Checklist::with('kid')->select('id', 'level', 'situation', 'kid_id', 'created_at');
             $data->where('created_by', '=', auth()->user()->id);
         }
+        */
+
+        $data = Checklist::getChecklists();
 
         return Datatables::of($data)
             ->addColumn('action', function ($data) {
-                if (request()->user()->can('checklists.update') || request()->user()->can('checklists.store')) {
+                $user = request()->user();
 
-                    $html = '<div class="dropdown">
-                      <button class="btn btn-sm btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
-                                            <i class="bi bi-gear"></i>
-                                        </button>
-                      <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
-                        <li><a class="dropdown-item" href="'.route('checklists.edit', $data->id).'"><i class="bi bi-pencil"></i> Editar</a></li>
-                        <li><a class="dropdown-item" href="'.route('checklists.fill', $data->id).'"><i class="bi bi-check2-square"></i> Avaliação</a></li>
-                      </ul>
-                    </div>';
+                $html = '<div class="dropdown">
+                    <button class="btn btn-sm btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
+                        <i class="bi bi-gear"></i>
+                    </button>
+                    <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">';
 
-                    return $html;
-                }
+                    // Adiciona o botão de visualizar se o usuário tiver permissão de visualizar checklists
+                    if ($user->can('view checklists')) {
+                        $html .= '<li><a class="dropdown-item" href="'.route('checklists.show', $data->id).'">
+                                    <i class="bi bi-eye"></i> Visualizar
+                                </a></li>';
+                    }
+
+                    // Adiciona o botão de editar se o usuário tiver permissão de editar checklists
+                    if ($user->can('edit checklists')) {
+                        $html .= '<li><a class="dropdown-item" href="'.route('checklists.edit', $data->id).'">
+                                    <i class="bi bi-pencil"></i> Editar
+                                </a></li>';
+                    }
+
+                    // Adiciona o botão de avaliação se o usuário tiver permissão de preencher checklists (fill)
+                    if ($user->can('fill checklists')) {
+                        $html .= '<li><a class="dropdown-item" href="'.route('checklists.fill', $data->id).'">
+                                    <i class="bi bi-check2-square"></i> Avaliação
+                                </a></li>';
+                    }
+
+                $html .= '</ul></div>';
+
+                return $html;
             })
             ->editColumn('kid_id', function ($data) {
                 return $data->kid->name;
@@ -66,6 +90,8 @@ class ChecklistController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Checklist::class);
+
         $message = label_case('Create Checklist ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
         Log::info($message);
         $kids = Kid::all('id', 'name');
@@ -75,6 +101,8 @@ class ChecklistController extends Controller
 
     public function store(ChecklistRequest $request)
     {
+        $this->authorize('create', Checklist::class);
+
         try {
             $message = label_case('Store Checklists '.self::MSG_UPDATE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::info($message);
@@ -88,6 +116,7 @@ class ChecklistController extends Controller
             // Plane
             $plane = Plane::create([
                 'kid_id' => $request->kid_id,
+                'checklist_id' => $checklist->id,
                 'created_by' => Auth::id(),
             ]);
 
@@ -110,8 +139,8 @@ class ChecklistController extends Controller
             flash(self::MSG_UPDATE_SUCCESS)->success();
 
             return redirect()->route('checklists.index');
-        } catch (Exception $e) {
-
+        } catch (\Exception $e) {
+            dd($e->getMessage());
             $message = label_case('Create Checklists '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::error($message);
 
@@ -123,20 +152,40 @@ class ChecklistController extends Controller
 
     public function show($id)
     {
-        dd('show');
+        $checklist = Checklist::findOrFail($id);
+        $this->authorize('view', $checklist);
+
+        try {
+            $message = label_case('Edit Checklist ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            Log::info($message);
+
+            return view('checklists.show', [
+                'checklist' => $checklist,
+            ]);
+        } catch (\Exception $e) {
+            dd($e->getMessage());
+            $message = label_case('Update Checklist '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            Log::error($message);
+
+            flash(self::MSG_UPDATE_ERROR)->warning();
+            return redirect()->back();
+        }
     }
 
     public function edit($id)
     {
+        $this->authorize('edit checklists');
+
         try {
-            $checklist = Checklist::findOrFail($id);
             $message = label_case('Edit Checklist ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::info($message);
+
+            $checklist = Checklist::findOrFail($id);
 
             return view('checklists.edit', [
                 'checklist' => $checklist,
             ]);
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             flash(self::MSG_UPDATE_ERROR)->warning();
             $message = label_case('Update Checklist '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::error($message);
@@ -147,19 +196,21 @@ class ChecklistController extends Controller
 
     public function update(ChecklistRequest $request, $id)
     {
+        $checklist = Checklist::findOrFail($id);
+        $this->authorize('update', $checklist);
+
         try {
             $message = label_case('Update Checklists '.self::MSG_UPDATE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::info($message);
 
             $data = $request->all();
             $data['updated_by'] = Auth::id();
-            $checklist = Checklist::findOrFail($id);
             $checklist->update($data);
 
             flash(self::MSG_UPDATE_SUCCESS)->success();
 
             return redirect()->route('checklists.index');
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
 
             $message = label_case('Update Checklists '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::error($message);
@@ -172,25 +223,38 @@ class ChecklistController extends Controller
 
     public function destroy($id)
     {
+        $checklist = Checklist::findOrFail($id);
+        $this->authorize('delete', $checklist);
+
         try {
             $message = label_case('Destroy Checklist '.self::MSG_DELETE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::info($message);
+            if ($checklist->planes()->exists()) {
+                foreach ($checklist->planes as $plane) {
+                    $plane->deleted_by = Auth::id(); // Atribui o usuário que deletou
+                    $plane->save(); // Salva as alterações no banco de dados
+                    $plane->delete(); // Realiza a exclusão (soft delete, se for o caso)
+                }
+            }
 
-            $checklist = Checklist::findOrFail($id);
+            // Exclui o Checklist
             $checklist->deleted_by = Auth::id();
-            $checklist->update();
+            $checklist->save();
             $checklist->delete();
+
             flash(self::MSG_DELETE_SUCCESS)->success();
 
             return redirect()->route('checklists.index');
-        } catch (Exception $e) {
-            flash(self::MSG_NOT_FOUND)->warning();
-            $message = label_case('Destroy Checkilist '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+        } catch (\Exception $e) {
+            dd($e->getMessage());
+            $message = label_case('Destroy Checklist '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::error($message);
 
+            flash(self::MSG_NOT_FOUND)->warning();
             return redirect()->back();
         }
     }
+
 
     public function fill($id)
     {
@@ -204,18 +268,18 @@ class ChecklistController extends Controller
                 'created_at' => $checklist->created_at->format('d/m/Y').' Cod. '.$id,
                 'kid' => $checklist->kid,
             ];
-
             return view('checklists.fill', $data);
-        } catch (Exception $e) {
-            flash(self::MSG_NOT_FOUND)->warning();
+        } catch (\Exception $e) {
             $message = label_case('Fill Checklist '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
             Log::error($message);
 
-            return redirect()->back();
+            dd($e->getMessage());
+            flash(self::MSG_NOT_FOUND)->warning();
+            //return redirect()->back();
         }
     }
 
-    public function register(Request $request)
+    /*public function register(Request $request)
     {
         $data = $request->all();
         $checklistRegister = ChecklistRegister::where('checklist_id', $request->checklist_id)->where('competence_description_id', $request->competence_description_id);
@@ -226,7 +290,7 @@ class ChecklistController extends Controller
         } else {
             $checklistRegister->create($data);
         }
-    }
+    }*/
 
     public function chart($id)
     {
