+++ app/Http/Controllers/KidsController.php
@@ -9,15 +9,18 @@ use App\Models\Plane;
 use App\Models\Responsible;
 use App\Models\User;
 use App\Util\MyPdf;
+use Auth;
 use Exception;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Yajra\DataTables\DataTables;
+use Spatie\Permission\Models\Role as SpatieRole;
 
 class KidsController extends Controller
 {
     public function index()
     {
-        $message = label_case('Index Kids ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+        $message = label_case('Index Kids ') . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
         Log::debug($message);
 
         return view('kids.index');
@@ -30,14 +33,14 @@ class KidsController extends Controller
         return Datatables::of($data)
             ->addColumn('action', function ($data) {
 
-                if (request()->user()->can('kids.update') || request()->user()->can('kids.store')) {
+                if (request()->user()->can('edit kids')) {
 
                     $html = '<div class="dropdown">';
                     $html .= '<button class="btn btn-sm btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false"><i class="bi bi-gear"></i></button><ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">';
-                    $html .= '<li><a class="dropdown-item" href="'.route('kids.edit', $data->id).'"><i class="bi bi-pencil"></i> Editar</a></li>';
+                    $html .= '<li><a class="dropdown-item" href="' . route('kids.edit', $data->id) . '"><i class="bi bi-pencil"></i> Editar</a></li>';
 
                     if ($data->checklists()->count() > 0) {
-                        $html .= '<li><a class="dropdown-item" href="'.route('kids.show', $data->id).'"><i class="bi bi-check2-square"></i> Checklist</a></li>';
+                        $html .= '<li><a class="dropdown-item" href="' . route('kids.show', $data->id) . '"><i class="bi bi-check2-square"></i> Checklist</a></li>';
                     }
 
                     $html .= '</ul></div>';
@@ -52,47 +55,64 @@ class KidsController extends Controller
                 return $data->birth_date;
             })
             ->editColumn('checklists', function ($data) {
-                return '<span class="badge bg-success"><i class="bi bi-check"></i> '.$data->checklists->count().' Checklist(s) </span>';
+                return '<span class="badge bg-success"><i class="bi bi-check"></i> ' . $data->checklists->count() . ' Checklist(s) </span>';
             })
-            ->editColumn('user_id', function ($data) {
-                if ($data->user_id) {
-                    return $data->user->name;
-                }
+            ->editColumn('profession_id', function ($data) {
+                // Exibe o nome do professional ou 'Não atribuído' caso não tenha um professional
+                return $data->professional ? $data->professional->name : 'Não atribuído';
             })
             ->editColumn('responsible_id', function ($data) {
-                if ($data->responsible_id) {
-                    return $data->responsible->name;
-                }
+                // Exibe o nome do responsável ou 'Não atribuído' caso não tenha um responsável
+                return $data->responsible ? $data->responsible->name : 'Não atribuído';
             })
-            ->rawColumns(['name', 'checklists', 'user_id', 'responsible_id', 'action'])
-            ->orderColumns(['id'], '-:column $1')
+            ->rawColumns(['name', 'checklists', 'responsible', 'action'])
+            //->orderColumns(['id'], '-:column $1')
             ->make(true);
     }
 
     public function create()
     {
-        $message = label_case('Create Kids').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+        $this->authorize('create', Kid::class);
+
+        $message = label_case('Create Kids') . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
         Log::info($message);
 
-        $users = User::scopeListUsers();
-        $responsibles = Responsible::all();
-
-        return view('kids.create', compact('users', 'responsibles'));
+        return view('kids.create');
     }
 
     public function store(KidRequest $request)
     {
+        $this->authorize('create', Kid::class);
+
+        DB::beginTransaction();
         try {
-            $message = label_case('Store Kids '.self::MSG_CREATE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Store Kids ' . self::MSG_CREATE_SUCCESS) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::info($message);
 
-            Kid::create($request->all());
+
+            $kidData = [
+                'name' => $request->name,
+                'birth_date' => $request->birth_date,
+                'created_by' => auth()->user()->id,
+            ];
+
+            if (Auth::user()->hasRole('professional')) {
+                $kidData['profession_id'] = Auth::user()->id;
+            }
+
+            $kid = Kid::forProfessional()->create($kidData);
+
+            Log::info('Kid created: ' . $kid->id . ' created by: ' . auth()->user()->id);
+
             flash(self::MSG_CREATE_SUCCESS)->success();
+
+            DB::commit();
 
             return redirect()->route('kids.index');
         } catch (Exception $e) {
+            DB::rollBack();
             flash(self::MSG_CREATE_ERROR)->warning();
-            $message = label_case('Store Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Store Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
 
             return redirect()->route('kids.index');
@@ -101,9 +121,13 @@ class KidsController extends Controller
 
     public function show(Kid $kid)
     {
+        $this->authorize('view', $kid);
+
         try {
-            $message = label_case('Show Kids ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
-            Log::info($message);
+            Log::info('show', [
+                'user' => auth()->user()->name,
+                'id' => auth()->user()->id,
+            ]);
 
             if ($kid->checklists()->count() === 0) {
                 flash(self::MSG_NOT_FOUND_CHECKLIST_USER)->warning();
@@ -113,10 +137,10 @@ class KidsController extends Controller
 
             $checklists = $kid->checklists()->orderBy('created_at', 'DESC')->get();
             $kid->months = $kid->months;
-            $kid->profession = $kid->user->name;
+            //$kid->profession = $kid->user->name;
             $data = [
                 'kid' => $kid,
-                'profession' => $kid->user->name,
+                'profession' => $kid->professional->name,
                 'checklists' => $checklists,
                 'checklist_id' => $checklists[0]->id,
                 'level' => $checklists[0]->level,
@@ -126,27 +150,49 @@ class KidsController extends Controller
 
             return view('kids.show', $data);
         } catch (Exception $e) {
-            flash(self::MSG_NOT_FOUND)->warning();
-            $message = label_case('Show Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Error show kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
 
+            flash(self::MSG_NOT_FOUND)->warning();
             return redirect()->back();
         }
     }
 
     public function edit(Kid $kid)
     {
+        $this->authorize('update', $kid);
+
         try {
-            $message = label_case('Edit Kids ').' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Edit Kids ') . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::info($message);
 
-            $users = User::scopeListUsers();
-            $responsibles = Responsible::all();
+            // Verificando se o papel 'pais' existe, caso contrário lançar exceção
+            $parentRole = SpatieRole::where('name', 'pais')->first();
+            if (!$parentRole) {
+                throw new Exception("O papel 'pais' não existe no sistema.");
+            }
 
-            return view('kids.edit', compact('kid', 'users', 'responsibles'));
+            // Verificando se o papel 'professional' existe, caso contrário lançar exceção
+            $professionalRole = SpatieRole::where('name', 'professional')->first();
+            if (!$professionalRole) {
+                throw new Exception("O papel 'professional' não existe no sistema.");
+            }
+
+            // Buscando usuários com o papel 'pais'
+            $responsibles = User::whereHas('roles', function ($query) {
+                $query->where('name', 'pais');
+            })->get();
+
+            // Buscando usuários com o papel 'professional'
+            $professions = User::whereHas('roles', function ($query) {
+                $query->where('name', 'professional');
+            })->get();
+
+
+            return view('kids.edit', compact('kid', 'responsibles', 'professions'));
         } catch (Exception $e) {
-            flash(self::MSG_NOT_FOUND)->warning();
-            $message = label_case('Update Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            flash($e->getMessage())->warning();
+            $message = label_case('Update Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
 
             return redirect()->back();
@@ -155,40 +201,63 @@ class KidsController extends Controller
 
     public function update(KidRequest $request, Kid $kid)
     {
+        // Verifica se o usuário está autorizado a atualizar o registro
+        $this->authorize('update', $kid);
+
         try {
-            $message = label_case('Update Kids '.self::MSG_UPDATE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            // Loga a tentativa de atualização
+            $message = label_case('Update Kids ' . self::MSG_UPDATE_SUCCESS) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::info($message);
 
+            // Coleta os dados do request
             $data = $request->all();
 
-            $kid->update($data);
-            SendKidUpdateJob::dispatch($kid)->onQueue('emails');
+            // Removida a verificação de papel 'isProfessional'
+            // O tratamento para usuários profissionais será feito futuramente
 
+            // Atualiza os dados da criança
+            $kid->update($data);
+
+            // Opcional: Disparar job de atualização de criança
+            // SendKidUpdateJob::dispatch($kid)->onQueue('emails');
+
+            // Mensagem de sucesso
             flash(self::MSG_UPDATE_SUCCESS)->success();
 
-            return redirect()->route('kids.index');
+            // Redireciona para a página de edição
+            return redirect()->route('kids.edit', $kid->id);
         } catch (Exception $e) {
-            //flash(self::MSG_UPDATE_ERROR)->warning();
-            echo $message = label_case('Update Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
-            //Log::error($message);
+            // Loga o erro em caso de falha
+            $message = label_case('Update Kids Error' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
+            Log::error($message);
 
-            //return redirect()->back();
+            // Mensagem de erro
+            flash(self::MSG_UPDATE_ERROR)->warning();
+
+            // Redireciona de volta para a página anterior
+            return redirect()->back();
         }
     }
 
+
     public function destroy(Kid $kid)
     {
+        $this->authorize('delete', $kid);
         try {
-            $message = label_case('Destroy Kids '.self::MSG_DELETE_SUCCESS).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+            $message = label_case('Destroy Kids ' . self::MSG_DELETE_SUCCESS) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::info($message);
+
+
             $kid->delete();
             flash(self::MSG_DELETE_SUCCESS)->success();
 
             return redirect()->route('kids.index');
         } catch (Exception $e) {
-            flash(self::MSG_NOT_FOUND)->warning();
-            $message = label_case('Destroy Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
+
+            $message = label_case('Destroy Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
             Log::error($message);
+
+            flash(self::MSG_NOT_FOUND)->warning();
 
             return redirect()->back();
         }
@@ -201,7 +270,7 @@ class KidsController extends Controller
             $kid_id = $plane->kid()->first()->id;
             $kid = Kid::findOrFail($kid_id);
             $nameKid = $plane->kid()->first()->name;
-            $therapist = $kid->user()->first()->name;
+            $therapist = $kid->professional->name;
             $date = $plane->first()->created_at;
             $arr = [];
 
@@ -239,13 +308,13 @@ class KidsController extends Controller
 
                     $pdf->Ln(5);
                     $pdf->SetFont('helvetica', 'B', 10);
-                    $txt = $competence->level_id.$v['domain']->initial.$competence->code.' - '.$competence->description;
+                    $txt = $competence->level_id . $v['domain']->initial . $competence->code . ' - ' . $competence->description;
                     $pdf->Ln(5);
                     $pdf->Write(0, $txt, '', 0, 'L', true);
 
                     $pdf->Ln(1);
                     $pdf->SetFont('helvetica', 'I', 8);
-                    $pdf->Write(0, '"'.$competence->description_detail.'"', '', 0, 'L', true);
+                    $pdf->Write(0, '"' . $competence->description_detail . '"', '', 0, 'L', true);
 
                     $pdf->Ln(4);
                     $pdf->SetFont('helvetica', '', 9);
@@ -258,13 +327,19 @@ class KidsController extends Controller
                 //                }
             }
 
-            $pdf->Output($nameKid.'_'.$date->format('dmY').'_'.$plane->id.'.pdf', 'I');
+            $pdf->Output($nameKid . '_' . $date->format('dmY') . '_' . $plane->id . '.pdf', 'I');
         } catch (Exception $e) {
-            flash(self::MSG_NOT_FOUND)->warning();
-            $message = label_case('Plane Kids '.$e->getMessage()).' | User:'.auth()->user()->name.'(ID:'.auth()->user()->id.')';
-            Log::error($message);
+            $message = label_case('Plane Kids ' . $e->getMessage()) . ' | User:' . auth()->user()->name . '(ID:' . auth()->user()->id . ')';
+            Log::error('Exibe Plano Erro', [
+                'message' => $e->getMessage(),
+                'file' => $e->getFile(),
+                'line' => $e->getLine(),
+            ]);
+            dd($e->getMessage());
 
-            return redirect()->route('kids.index');
+            flash(self::MSG_NOT_FOUND)->warning();
+
+            //return redirect()->route('kids.index');
         }
     }
 
@@ -292,9 +367,9 @@ class KidsController extends Controller
         $pdf->setViewerPreferences($preferences);
         $pdf->AddPage();
         $pdf->SetFont('helvetica', '', 18);
-        $pdf->Cell(0, 75, 'PLANO DE INTERVENÇÃO N.: '.$plane_id, 0, 1, 'C');
+        $pdf->Cell(0, 75, 'PLANO DE INTERVENÇÃO N.: ' . $plane_id, 0, 1, 'C');
         $pdf->SetFont('helvetica', '', 12);
-        $txt = 'Terapeuta: '.$therapist.',  Data: '.$date;
+        $txt = 'Terapeuta: ' . $therapist . ',  Data: ' . $date;
         $pdf->Cell(0, 2, $txt, 0, 1, 'C');
         $pdf->Ln(15);
         $pdf->SetFont('helvetica', '', 18);
